requirements.txt
discord.py==2.3.2
tweety-ns==1.0.9.2
python-dotenv==1.0.0
pyyaml==6.0.1
aiosqlite==0.18.0